Move set_file_descriptor_limit to common/util.cc
