+++ tools/replay/util.cc
@@ -320,17 +320,3 @@ std::string sha256(const std::string &str) {
   SHA256_Final(hash, &sha256);
   return util::hexdump(hash, SHA256_DIGEST_LENGTH);
 }
-
-int setFileDescriptorLimit(uint64_t limit_val) {
-  struct rlimit limit;
-  int status;
-
-  if ((status = getrlimit(RLIMIT_NOFILE, &limit)) < 0)
-    return status;
-
-  limit.rlim_cur = limit_val;
-  if ((status = setrlimit(RLIMIT_NOFILE, &limit)) < 0)
-    return status;
-
-  return 0;
-}
